All remaining initial bugs of Kitchensink is resolved.
