+++ packages/ext-web-components-kitchensink/src/view/d3/hierarchy/orgchart/OrgChartComponent.html
@@ -1,8 +1,13 @@
-<ext-panel  shadow="true" layout="fit" title="Organization Chart" fitToParent="true" height="100%">
-  <ext-d3-horizontal-tree
-    onready="orgchart.horizontaltreeReady(event)" 
-    nodeSize="[200, 100]"
-    imagePath="resources/images/org-chart/"
+<ext-panel
+  shadow="true"
+  layout="fit"
+  title="Organization Chart"
+  fitToParent="true"
+  height="100%"
+>
+  <ext-panel
+    onready="orgchart.horizontalTreePanelReady(event)"
     height="100%"
-  ></ext-d3-horizontal-tree>
+  >
+  </ext-panel>
 </ext-panel>
